requirements.txt
asgiref==3.3.4
certifi==2020.12.5
chardet==4.0.0
dj-database-url==0.5.0
Django==3.2
django-heroku==0.3.1
django-mysql==3.11.1
gunicorn==20.1.0
idna==2.10
Pillow==8.2.0
psycopg2==2.8.6
PyJWT==1.7.1
python-decouple==3.4
pytz==2021.1
requests==2.25.1
six==1.15.0
sqlparse==0.4.1
twilio==6.56.0
urllib3==1.26.4
whitenoise==5.2.0
